warp/analyzer/do_frame_transfer.py
import numpy as np
from typing import Dict, Any, Union

def do_frame_transfer(metric: Dict[str, Any], energy_tensor: Dict[str, Any], frame: str, try_gpu: int = 0) -> Dict[str, Any]:
    """
    Transforms the energy tensor into selected frames.

    Args:
        metric (dict): Metric struct.
        energy_tensor (dict): Energy struct.
        frame (str): Frame to transform the tensor to. Only 'Eulerian' is supported.
        try_gpu (int, optional): A flag on whether or not to use GPU computation (0=no, 1=yes). Default is 0.

    Returns:
        dict: Transformed energy struct.

    Raises:
        ValueError: If metric or energy tensor is not verified.
        ValueError: If an unsupported frame is provided.
    """
    transformed_energy_tensor = energy_tensor.copy()
    transformed_energy_tensor['tensor'] = [[None for _ in range(4)] for _ in range(4)]

    if not verify_tensor(metric):
        raise ValueError("Metric is not verified. Please verify metric using verify_tensor(metric).")
    if not verify_tensor(energy_tensor):
        raise ValueError("Stress-energy is not verified. Please verify Stress-energy tensor using verify_tensor(energy_tensor).")

    if frame.lower() == "eulerian" and not (energy_tensor.get('frame') and energy_tensor['frame'].lower() == 'eulerian'):
        # Convert to covariant (lower) index
        energy_tensor = change_tensor_index(energy_tensor, "covariant", metric)

        # Convert from cell to array
        array_energy_tensor = tensor_cell_to_array(energy_tensor, try_gpu)
        array_metric_tensor = tensor_cell_to_array(metric, try_gpu)

        # Do transformations at each point in space
        M = get_eulerian_transformation_matrix(array_metric_tensor, metric['coords'])
        M = np.transpose(M, (2, 3, 0, 1))
        array_energy_tensor = np.transpose(array_energy_tensor, (2, 3, 0, 1))

        transformed_temp_tensor = np.einsum('...ij,...jk,...kl->...il', M, array_energy_tensor, M)

        # Convert array tensor into cell format
        z = transformed_temp_tensor.shape
        for i in range(4):
            for j in range(4):
                transformed_energy_tensor['tensor'][i][j] = transformed_temp_tensor[i, j].reshape(z[2:], order='F')

        # Transform to contravariant T^{0, i} = -T_{0, i}
        for i in range(1, 4):
            transformed_energy_tensor['tensor'][0][i] = -transformed_energy_tensor['tensor'][0][i]
            transformed_energy_tensor['tensor'][i][0] = -transformed_energy_tensor['tensor'][i][0]

        # Update the tensor metadata
        transformed_energy_tensor['frame'] = "Eulerian"
        transformed_energy_tensor['index'] = "contravariant"
    else:
        raise ValueError("Unsupported frame or frame already set to 'Eulerian'.")

    return transformed_energy_tensor

def verify_tensor(tensor: Dict[str, Any]) -> bool:
    """
    Verify the tensor format.

    Args:
        tensor (dict): Tensor to verify.

    Returns:
        bool: True if tensor is valid, False otherwise.
    """
    required_keys = {'type', 'index', 'tensor'}
    if not required_keys.issubset(tensor.keys()):
        return False
    if tensor['type'] not in {'metric', 'energy'}:
        return False
    if not isinstance(tensor['tensor'], np.ndarray):
        return False
    return True

def change_tensor_index(tensor: Dict[str, Any], index: str, metric: Dict[str, Any]) -> Dict[str, Any]:
    """
    Change the index of the tensor. Placeholder for the actual implementation.
    """
    return tensor  # Placeholder implementation

def tensor_cell_to_array(tensor: Dict[str, Any], try_gpu: int) -> np.ndarray:
    """
    Convert a tensor from cell format to array format. Placeholder for actual conversion logic.
    """
    return np.random.rand(4, 4, 4, 4)  # Placeholder implementation

def get_eulerian_transformation_matrix(metric_tensor: np.ndarray, coords: Any) -> np.ndarray:
    """
    Get the Eulerian transformation matrix. Placeholder for the actual implementation.
    """
    return np.random.rand(4, 4, 4, 4)  # Placeholder implementation
